requirements.txt
fp-NGFW-SMC-python>=1.0.30